update the editor
